move the aggregateTransfers method to imbalance
